requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
jinja2>=3.1.3
python-multipart>=0.0.9
apscheduler>=3.10.4
numpy>=1.24.0
websocket-client>=1.7.0
